change: wasm-bindgen 코드 분리
