+++ src/functions/mod.rs
@@ -1,3 +1,2 @@
 pub mod random;
-pub mod render;
 pub mod valid_mino;
