+++ src/wasm_bindgen/mod.rs
@@ -1,0 +1,2 @@
+pub mod draw;
+pub mod render_loop;
